+++ legacy_packages/chains/chains/1802203764.ts
@@ -7,6 +7,11 @@ export default {
       "name": "Kakarot Scan",
       "url": "https://sepolia.kakarotscan.org",
       "standard": "EIP3091"
+    },
+    {
+      "name": "Kakarot Explorer",
+      "url": "https://sepolia-explorer.kakarot.org",
+      "standard": "EIP3091"
     }
   ],
   "faucets": [],