+++ legacy_packages/chains/chains/88882.ts
@@ -4,12 +4,19 @@ export default {
   "chainId": 88882,
   "explorers": [
     {
+      "name": "spicy-explorer",
+      "url": "https://testnet.chiliscan.com",
+      "standard": "EIP3091"
+    },
+    {
       "name": "Spicy Explorer",
       "url": "http://spicy-explorer.chiliz.com/",
       "standard": "none"
     }
   ],
   "faucets": [
+    "https://spicy-faucet.chiliz.com",
+    "https://tatum.io/faucets/chiliz",
     "https://spicy-faucet.chiliz.com/"
   ],
   "features": [],
@@ -30,9 +37,11 @@ export default {
   "redFlags": [],
   "rpc": [
     "https://88882.rpc.thirdweb.com/${THIRDWEB_API_KEY}",
-    "https://spicy-rpc.chiliz.com/"
+    "https://spicy-rpc.chiliz.com/",
+    "https://spicy-rpc.chiliz.com"
   ],
   "shortName": "Spicy",
+  "slip44": 1,
   "slug": "spicy-chain",
   "testnet": true
 } as const satisfies Chain;