+++ legacy_packages/chains/chains/8898.ts
@@ -6,12 +6,24 @@ export default {
     {
       "name": "mmtscan",
       "url": "https://mmtscan.io",
-      "standard": "EIP3091"
+      "standard": "EIP3091",
+      "icon": {
+        "url": "ipfs://QmaF5gi2CbDKsJ2UchNkjBqmWjv8JEDP3vePBmxeUHiaK4",
+        "width": 250,
+        "height": 250,
+        "format": "png"
+      }
     }
   ],
   "faucets": [
     "https://faucet.mmtscan.io/"
   ],
+  "icon": {
+    "url": "ipfs://QmaF5gi2CbDKsJ2UchNkjBqmWjv8JEDP3vePBmxeUHiaK4",
+    "width": 250,
+    "height": 250,
+    "format": "png"
+  },
   "infoURL": "https://mmtchain.io/",
   "name": "Mammoth Mainnet",
   "nativeCurrency": {