+++ packages/thirdweb/src/react/web/ui/ConnectWallet/constants.ts
@@ -12,7 +12,7 @@ export const modalMaxWidthWide = `${wideModalWidth}px`;
 export const wideModalScreenThreshold = wideModalWidth + 40;
 
 export const wideModalMaxHeight = "570px";
-export const compactModalMaxHeight = "650px";
+export const compactModalMaxHeight = "660px";
 
 export const defaultTheme = "dark";